Added new background, added grass class idea.
New background and grass.jpg uploaded to resources. The way the background behaves makes me want to do the level design a different way and just have a different class for every type of block and have all the blocks be 32,32. Working now on creating a Grass class, Wall class, etc.